+++ src/webpack/webpack.defaults.js
@@ -37,6 +37,7 @@ module.exports = config => ({
             // fix for broken RxJS requiring by webpack
             // TODO: remove once fixed in webpack
             rx: 'rx/dist/rx.all.js',
+            'ecc-superagent': '@eccenca/superagent',
         },
     },
     node: {