+++ test/fixtures/bad-mdl/bad-mdl.jsx
@@ -1,0 +1,2 @@
+
+const BAD = <div className="mdl-button" />