+++ test/fixtures/bad-mdl/bad.js
@@ -1,0 +1,6 @@
+
+var foo = 'mdl-card';
+
+var foo = 'mdl-card';
+var foo = 'mdl-card';
+var foo = 'mdl-button';