+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ### Added
 - bad mdl patterns are now recognized when running gulp doctor
+- Alias `ecc-superagent` for `@eccenca/superagent`
 
 ## [3.10.0] 2017-08-18