+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
+## [3.11.0] 2017-09-27
+
+### Added
+- bad mdl patterns are now recognized when running gulp doctor
+- Alias `ecc-superagent` for `@eccenca/superagent`
+
 ## [3.10.0] 2017-08-18
 
 ### Changed